# public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'CAA — API Documentation'
  description: 'API REST de la plateforme CAA (Certificat Authenticité Artiste) — gestion des artistes, œuvres, certificats numériques et vérification publique.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: "Obtenez votre token via `POST /api/v1/auth/login`. Envoyez-le ensuite dans l'en-tête `Authorization: Bearer <token>`."
security:
  -
    default: []
paths:
...
  /api/v1/auth/login:
    post:
      summary: Connexion
      operationId: connexion
      description: 'Authentifie un utilisateur et retourne un token Sanctum.'
      parameters: []
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 3
                    email: awa@caa.sn
                    roles:
                      - artist
                  token: 2|6URd4kiuvxY5pHKFzSnpYbP0LM3XkLattT2rTfCze475f5bb
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 3
                      email:
                        type: string
                        example: awa@caa.sn
                      roles:
                        type: array
                        example:
                          - artist
                        items:
                          type: string
                  token:
                    type: string
                    example: 2|6URd4kiuvxY5pHKFzSnpYbP0LM3XkLattT2rTfCze475f5bb
        422:
          description: 'invalid credentials'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Identifiants invalides.'
                  errors:
                    email:
                      - 'Identifiants invalides.'
                properties:
                  message:
                    type: string
                    example: 'Identifiants invalides.'
                  errors:
                    type: object
                    properties:
                      email:
                        type: array
                        example:
                          - 'Identifiants invalides.'
                        items:
                          type: string
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "Email de l'utilisateur."
                  example: awa@caa.sn
                password:
                  type: string
                  description: "Mot de passe de l'utilisateur."
                  example: awapass
                device_name:
                  type: string
                  description: "Nom de l'appareil ou du client API."
                  example: scalar
                  nullable: true
              required:
                - email
                - password
      security: []
...